New additional FT_Property_Get() prototype which takes a ref int as parameter.
I believe the original code [1] for PropertyGet with a "out IntPtr" last
argument is wrong for value types like int. Whereas pointer types get the
IntPtr filled in with by having the pointer to a pre-existing structure
within Freetype, when FT_property_get() is called with &int as a final
argument, the caller is expected to provide the storage which the pointer
is pointed to. Hence "out IntPtr" will never work when FT_Property_Get
expects int* , as the storage pointed to by the pointer never existed.

This enhancement fixes:
https://github.com/Robmaister/SharpFont/issues/82
(I tested this - it works for my purpose - doing:
   int b;
   _lib.PropertyGet("truetype", "interpreter-version", ref b); // b is 35
   _lib.PropertySet("truetype", "interpreter-version", 40);
   _lib.PropertyGet("truetype", "interpreter-version", ref b); // b is now 40
)

Also note that [1] said "untested" - do not know if it is still true:

[1]
commit 0ad176fc9bbfe0ebe50722d4a16388c1dae964e1
Author: Robert Rouhani <robert.rouhani@gmail.com>
Date:   Fri Jan 25 21:46:15 2013 -0500

    Updated to FreeType 2.4.11, untested.